Update async_utils.py
with the previous solution, the wait function was not working when called behind a qt slot. So i have rewritten the asyncio.wait() so that we can pass the LOOP parameter which we couldn't do since python 3.10. Calling LOOP.run_until_complete() blocks when called behind a qt slot which I didn't see before.